Fix installer app
